Moved check_type_compat and return_family_type to nonh5utils
